Reference

ParallelMaximumClique.maximum_clique

ParallelMaximumClique.maximum_clique

ParallelMaximumClique.maximum_clique!

ParallelMaximumClique.normalize!

ParallelMaximumClique.normalize!

ParallelMaximumClique.maximum_clique! — Method

int K = maximum_clique!(ouput::Vector{Int32}, ei::Vector{Int32}, ej::Vector{Int32}; 
    verbose=0; 
    algorithm=0; 
    time_limit_secs=3600.0;
    remove_time_secs=4.0)

Inputs:

ei: the vector of edge sources

ej: the vector of edge destinations

Outputs:

K: the maximum clique size

output: the vector containing a list of nodes in the maximum clique (modified in-place)

Parameters:

verbose: set to positive integer to print additional debug statements

algorithm: choose between {0,1,2}

time_limit_secs: maximum compute time

remove_time_secs: maximum time before the graph gets reduced.

Notes:

for an edge to be valid, ei[k] > ej[k] for all k.

only the first K elements of output are modified.

ParallelMaximumClique.maximum_clique — Method

maximum_clique(A::SparseMatrixCSC; normalize=true, kwargs...)

returns a Vector{Int32} of the nodes in the maximum clique of g, where A is the adjacency matrix of g.

all kwargs are passed to maximum_clique!

ParallelMaximumClique.maximum_clique — Method

maximum_clique(g::G; kwargs...) where {T, G<:AbstractGraph{T}}

returns a Vector{Int32} of the nodes of g that are in the maximum clique in g

all kwargs are passed to maximum_clique!

ParallelMaximumClique.normalize! — Method

normalize!(A<:AbstractMatrix)

replaces each non-zero element of A with a 1.

ParallelMaximumClique.normalize! — Method

normalize!(A::SparseMatrixCSC)

replaces each non-zero element of A with a 1.